+++ js/router/store/reducers/devices.js
@@ -60,4 +60,3 @@ export default (state = {}, action = {}) => {
       return state;
   }
 };
-